Add support for tags and searching them
This adds support for identifying tags within a task.
For the desc box, the tags are colored differently
There is also a "tagBucket" that collects all the tags in the file and allows for searching them.
The search adds support for searching for equality for a specific field/value.  This is what drives the tag search.