Fully integrates Drive support & adds x-editable
This finishes the integration between Drive, allowing a file to be stored on drive, downloaded locally, and then updated again with fresh changes.
Added a title to the task list which is linked to the TaskList title.  This allows for saving to work with unique-ish file names.
Added libraries to allow for editing the title with a single click.
The project specific task list is now available on Google Drive and will be tracked there instead of in git.  It has not been removed yet, but it will be soon.