Support task completion
This adds support for task completion along with a new set of action buttons.
This handles some of the places where complete tasks should/should not be considered.  Modified the duration to not account for complete tasks.  The completion should save with tasks.
Depending on the setting, complete tasks may just show with a strike through instead of being hidden.  This flag can not be edited currently.